#177 helpers file renamed to generate token
