+++ src/controllers/helpers/generate_token.js
